update add password validation into delete router

+++ api/routes/users.js
@@ -34,12 +34,27 @@ router.delete('/:id', async (req, res) => {
         const user = await User.findById(req.params.id);
         if(user) {
             try {
-                await Post.deleteMany({ username : user.username })
-                await User.findByIdAndDelete(req.params.id);
-                res.status(200).json("계정이 삭제되었습니다.")
+                console.log("비밀번호확인시작");
+                const validate = await bcrypt.compare(req.body.password, user.password);
+                if(validate) {
+                    const {password, ...나머지정보들} = user._doc;
+                    res.status(200).json(나머지정보들);
+                    console.log("사용자 인증 성공");
+                    try {
+                        await Post.deleteMany({ username : user.username })
+                        await User.findByIdAndDelete(req.params.id);
+                        res.status(200).json("계정이 삭제되었습니다.")
+                    } catch (err) {
+                        res.status(500).json(err);
+                    }
+                } else {
+                    res.status(400).json("비밀번호가 틀렸습니다.(비밀번호가 틀림)");
+                }
             } catch (err) {
+                console.log("비번확인중서버에러");
                 res.status(500).json(err);
             }
+            
         } else {
             res.status(404).json("사용자를 찾을 수 없습니다.")
         }
